Moved command manager and plugin manager into the main executable
tidied up the project slightly, by moving stuff around, and changing a
few namespaces.
