+++ AquaConsole/AquaConsole.csproj
@@ -76,6 +76,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Managers\CommandManager.cs" />
     <Compile Include="Commands\cap.cs" />
     <Compile Include="Commands\cd.cs" />
     <Compile Include="Commands\clear screen.cs" />
@@ -86,6 +87,8 @@
     <Compile Include="Commands\quit.cs" />
     <Compile Include="Commands\title.cs" />
     <Compile Include="Commands\version.cs" />
+    <Compile Include="Managers\DateEvents.cs" />
+    <Compile Include="Managers\PluginManager.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
@@ -112,7 +115,6 @@
     </BootstrapperPackage>
   </ItemGroup>
   <ItemGroup>
-    <Content Include="icon.ico" />
     <None Include="Resources\Icon.ico" />
   </ItemGroup>
   <ItemGroup>
