+++ AquaConsole/Commands/title.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace HackTool.Commands
+namespace AquaConsole.Commands
 {
     class title : ICommand
     {
